refactor(tests): refactor test folder structure
This branch implements a restructuring of the test suite.

Rationale: Over-modularization of test parameter definitions clutters up the test
suite and leads to lengthy/repetitive import statements.

For now, the structure of the test suite should only differentiate
1. unit tests and 2. all other tests. Additional test categories may
be introduced at a later point.

The test category should be indicated in the test module doc string.

Closes #100
